requirements.txt
greenlet==1.1.2
mysql-connector-python==8.0.29
protobuf==3.20.1
PyMySQL==1.0.2
